Refactor and change model to shared
