+++ Frontend/Components/GemStats.razor
@@ -1,4 +1,5 @@
-﻿@using global::Model
+﻿@using Shared.Entity
+@using PoEGamblingHelper3.Components.Model
 <div class="card mb-5" style="width: 45rem;">
     <div class="card-body">
         <h3 class="card-title">
