+++ Frontend/Components/GemStats.razor.css
