+++ Frontend/Components/InfiniteScrolling.razor
@@ -1,4 +1,4 @@
-﻿@using PoEGamblingHelper3.Shared.Service
+﻿@using PoEGamblingHelper3.Components.Service
 @implements IDisposable
 
 @code {
