+++ Frontend/Frontend.csproj
@@ -10,11 +10,11 @@
 
     <ItemGroup>
         <PackageReference Include="Blazored.LocalStorage" Version="4.3.0"/>
-        <PackageReference Include="Blazored.SessionStorage" Version="2.2.0"/>
-        <PackageReference Include="Blazored.Toast" Version="3.2.2"/>
+        <PackageReference Include="Blazored.SessionStorage" Version="2.3.0"/>
+        <PackageReference Include="Blazored.Toast" Version="4.0.0"/>
         <PackageReference Include="Blazored.Typeahead" Version="4.7.0"/>
-        <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="7.0.1"/>
-        <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="7.0.1" PrivateAssets="all"/>
+        <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly" Version="7.0.2"/>
+        <PackageReference Include="Microsoft.AspNetCore.Components.WebAssembly.DevServer" Version="7.0.2" PrivateAssets="all"/>
     </ItemGroup>
 
     <ItemGroup>
@@ -22,16 +22,16 @@
     </ItemGroup>
     
     <ItemGroup Condition="'$(Configuration)' == 'Debug'">
-        <ProjectReference Include="..\Model\Model.csproj"/>
+        <ProjectReference Include="..\Shared\Shared.csproj"/>
     </ItemGroup>
     <ItemGroup Condition="'$(Configuration)' == 'Release'">
-        <Reference Include="packages\Model">
-            <HintPath>packages\Model.dll</HintPath>
+        <Reference Include="packages\Shared">
+            <HintPath>packages\Shared.dll</HintPath>
         </Reference>
     </ItemGroup>
 
     <ItemGroup>
-        <Content Include="Shared\Filter.razor.cs">
+        <Content Include="Components\Filter.razor.cs">
             <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
         </Content>
     </ItemGroup>
