+++ Frontend/Program.cs
@@ -4,10 +4,8 @@ using Blazored.SessionStorage;
 using Blazored.Toast;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
-using Microsoft.JSInterop;
 using PoEGamblingHelper3;
-using PoEGamblingHelper3.Shared;
-using PoEGamblingHelper3.Shared.Service;
+using PoEGamblingHelper3.Components.Service;
 
 Thread.CurrentThread.CurrentCulture = CultureInfo.CreateSpecificCulture("en-US");
 CultureInfo.DefaultThreadCurrentCulture = CultureInfo.CreateSpecificCulture("en-US");
@@ -21,11 +19,11 @@ builder.Services.AddScoped(_ => new HttpClient
                                     BaseAddress = new Uri(builder.Configuration["BackendUrl"] ??
                                                           builder.HostEnvironment.BaseAddress + "/api/")
                                 });
-builder.Services.AddScoped<IGemService>(sp => new GemService(sp.GetService<HttpClient>()!));
-builder.Services.AddScoped<ITempleCostService>(sp => new TempleCostService(sp.GetService<HttpClient>()!));
-builder.Services.AddScoped<ICurrencyService>(sp => new CurrencyService(sp.GetService<HttpClient>()!));
-builder.Services.AddScoped<ILeagueService>(sp => new LeagueService(sp.GetService<HttpClient>()!));
-builder.Services.AddScoped<IScrollInfoService>(sp => new ScrollInfoService(sp.GetService<IJSRuntime>()!));
+builder.Services.AddScoped<IGemService, GemService>();
+builder.Services.AddScoped<ITempleCostService, TempleCostService>();
+builder.Services.AddScoped<ICurrencyService, CurrencyService>();
+builder.Services.AddScoped<ILeagueService, LeagueService>();
+builder.Services.AddScoped<IScrollInfoService, ScrollInfoService>();
 builder.Services.AddBlazoredToast();
 builder.Services.AddBlazoredSessionStorage();
 builder.Services.AddBlazoredLocalStorage();
