--- Frontend/Shared/JavaScript/Theme.cs
@@ -1,7 +1,0 @@
-﻿namespace PoEGamblingHelper3.Shared.JavaScript;
-
-public enum Theme
-{
-    Dark,
-    Light
-}
