--- Frontend/Shared/Service/ICurrencyService.cs
@@ -1,8 +1,0 @@
-﻿using Model;
-
-namespace PoEGamblingHelper3.Shared.Service;
-
-public interface ICurrencyService
-{
-    public Task<List<Currency>> GetAll();
-}
