--- Frontend/Shared/Service/ILeagueService.cs
@@ -1,8 +1,0 @@
-﻿using Model;
-
-namespace PoEGamblingHelper3.Shared.Service;
-
-public interface ILeagueService
-{
-    public Task<League> GetCurrent();
-}
