--- Frontend/Shared/Service/ITempleCostService.cs
@@ -1,8 +1,0 @@
-﻿using Model;
-
-namespace PoEGamblingHelper3.Shared.Service;
-
-public interface ITempleCostService
-{
-    public Task<TempleCost> Get();
-}
