+++ Frontend/_Imports.razor
@@ -7,8 +7,7 @@
 @using Microsoft.AspNetCore.Components.WebAssembly.Http
 @using Microsoft.JSInterop
 @using PoEGamblingHelper3
-@using PoEGamblingHelper3.Shared
+@using PoEGamblingHelper3.Components
 @using Blazored.Typeahead
 @using Blazored.Toast
 @using Blazored.Toast.Services
-@using Model
